+++ src/Taskling.SqlServer.Tests/Repositories/Given_RangeBlockRepository/When_GetLastDateRangeBlock.cs
@@ -20,7 +20,7 @@ public class When_GetLastDateRangeBlock : TestBase
     private readonly ILogger<When_GetLastDateRangeBlock> _logger;
     private readonly IRangeBlockRepository _rangeBlockRepository;
 
-    private readonly int _taskDefinitionId;
+    private readonly long _taskDefinitionId;
     private DateTime _baseDateTime;
 
     private long _block1;
@@ -28,7 +28,7 @@ public class When_GetLastDateRangeBlock : TestBase
     private long _block3;
     private long _block4;
     private long _block5;
-    private int _taskExecution1;
+    private long _taskExecution1;
 
     public When_GetLastDateRangeBlock(IBlocksHelper blocksHelper, IRangeBlockRepository rangeBlockRepository,
         IExecutionsHelper executionsHelper, IClientHelper clientHelper, ILogger<When_GetLastDateRangeBlock> logger,
