+++ src/Taskling.SqlServer/Models/BlockExecution.cs
@@ -3,7 +3,7 @@
 public class BlockExecution
 {
     public long BlockExecutionId { get; set; }
-    public int TaskExecutionId { get; set; }
+    public long TaskExecutionId { get; set; }
     public long BlockId { get; set; }
     public DateTime? StartedAt { get; set; }
     public DateTime? CompletedAt { get; set; }
