+++ src/Taskling.SqlServer/Models/TaskDefinition.cs
@@ -8,18 +8,18 @@ public class TaskDefinition
         TaskExecutions = new HashSet<TaskExecution>();
     }
 
-    public int TaskDefinitionId { get; set; }
+    public long TaskDefinitionId { get; set; }
     public string? ApplicationName { get; set; }
     public string? TaskName { get; set; }
     public DateTime? LastCleaned { get; set; }
     public string? ExecutionTokens { get; set; }
     public int UserCsStatus { get; set; }
-    public int? UserCsTaskExecutionId { get; set; }
+    public long? UserCsTaskExecutionId { get; set; }
     public string? UserCsQueue { get; set; }
     public int ClientCsStatus { get; set; }
-    public int? ClientCsTaskExecutionId { get; set; }
+    public long? ClientCsTaskExecutionId { get; set; }
     public string? ClientCsQueue { get; set; }
-    public int? HoldLockTaskExecutionId { get; set; }
+    public long? HoldLockTaskExecutionId { get; set; }
 
     public virtual ICollection<Block> Blocks { get; set; }
     public virtual ICollection<TaskExecution> TaskExecutions { get; set; }
