+++ src/Taskling.SqlServer/Models/TaskExecution.cs
@@ -8,8 +8,8 @@ public class TaskExecution
         TaskExecutionEvents = new HashSet<TaskExecutionEvent>();
     }
 
-    public int TaskExecutionId { get; set; }
-    public int TaskDefinitionId { get; set; }
+    public long TaskExecutionId { get; set; }
+    public long TaskDefinitionId { get; set; }
     public DateTime StartedAt { get; set; }
     public DateTime? CompletedAt { get; set; }
     public DateTime LastKeepAlive { get; set; }
