+++ src/Taskling.SqlServer/Models/TaskExecutionEvent.cs
@@ -3,7 +3,7 @@
 public class TaskExecutionEvent
 {
     public long TaskExecutionEventId { get; set; }
-    public int TaskExecutionId { get; set; }
+    public long TaskExecutionId { get; set; }
     public int EventType { get; set; }
     public string? Message { get; set; }
     public DateTime EventDateTime { get; set; }
