+++ src/Taskling.SqlServer/Tokens/Executions/TokenRequest.cs
@@ -5,7 +5,7 @@ namespace Taskling.SqlServer.Tokens.Executions;
 public class TokenRequest
 {
     public TaskId? TaskId { get; set; }
-    public int TaskDefinitionId { get; set; }
-    public int TaskExecutionId { get; set; }
+    public long TaskDefinitionId { get; set; }
+    public long TaskExecutionId { get; set; }
     public int ConcurrencyLimit { get; set; }
 }
