+++ src/Taskling.SqlServer/Tokens/ICommonTokenRepository.cs
@@ -4,9 +4,9 @@ namespace Taskling.SqlServer.Tokens;
 
 public interface ICommonTokenRepository
 {
-    Task AcquireRowLockAsync(int taskDefinitionId, int taskExecutionId, TasklingDbContext dbContext);
+    Task AcquireRowLockAsync(long taskDefinitionId, long taskExecutionId, TasklingDbContext dbContext);
 
-    Task<List<TaskExecutionState>> GetTaskExecutionStatesAsync(List<int> taskExecutionIds,
+    Task<List<TaskExecutionState>> GetTaskExecutionStatesAsync(List<long> taskExecutionIds,
         TasklingDbContext dbContext);
 
     bool HasExpired(TaskExecutionState taskExecutionState);
