+++ src/Taskling/CleanUp/ICleanUpService.cs
@@ -5,6 +5,6 @@ namespace Taskling.CleanUp;
 
 public interface ICleanUpService
 {
-    void CleanOldData(TaskId taskId, int taskExecutionId,
+    void CleanOldData(TaskId taskId, long taskExecutionId,
         ITaskConfigurationRepository taskConfigurationRepository);
 }
